# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ACCENTECX AI CARE Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /screenings:
    get:
      operationId: listScreenings
      tags: [screenings]
      summary: List screenings (optionally filter by child)
      parameters:
        - name: childId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of screenings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Screening"
    post:
      operationId: createScreening
      tags: [screenings]
      summary: Submit a new developmental screening
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
  /screenings/{id}:
    get:
      operationId: getScreening
      tags: [screenings]
      summary: Get a screening by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Screening detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"

  # ─── Appointments ────────────────────────────────────────────────────────────
...
  /therapy-session-notes:
    get:
      operationId: listTherapySessionNotes
      tags: [therapySessionNotes]
      summary: List therapy session notes
      parameters:
        - name: therapyPlanId
          in: query
          schema:
            type: integer
        - name: childId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Session notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TherapySessionNote"
    post:
      operationId: createTherapySessionNote
      tags: [therapySessionNotes]
      summary: Create a therapy session note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TherapySessionNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TherapySessionNote"
  /therapy-session-notes/{id}:
    patch:
      operationId: updateTherapySessionNote
      tags: [therapySessionNotes]
      summary: Update a therapy session note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TherapySessionNoteInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TherapySessionNote"
    delete:
      operationId: deleteTherapySessionNote
      tags: [therapySessionNotes]
      summary: Delete a therapy session note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Developmental Milestones ────────────────────────────────────────────────
...
components:
  schemas:
...
    Screening:
      type: object
      required: [id, childId, screeningType, status, riskLevel, createdAt]
      properties:
        id:
          type: integer
        childId:
          type: integer
        childName:
          type: ["string", "null"]
        screeningType:
          type: string
          enum: [parent_questionnaire, teacher_report, clinical_intake, behavioral_observation]
        status:
          type: string
          enum: [pending, in_progress, completed, reviewed]
        riskLevel:
          type: ["string", "null"]
          enum: [low, moderate, high, critical, null]
        communicationScore:
          type: ["integer", "null"]
        socialScore:
          type: ["integer", "null"]
        attentionScore:
          type: ["integer", "null"]
        motorScore:
          type: ["integer", "null"]
        emotionalScore:
          type: ["integer", "null"]
        behavioralClusters:
          type: ["string", "null"]
        referralRecommendations:
          type: ["string", "null"]
        clinicalSummary:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    ScreeningInput:
      type: object
      required: [childId, screeningType]
      properties:
        childId:
          type: integer
        screeningType:
          type: string
          enum: [parent_questionnaire, teacher_report, clinical_intake, behavioral_observation]
        communicationScore:
          type: integer
        socialScore:
          type: integer
        attentionScore:
          type: integer
        motorScore:
          type: integer
        emotionalScore:
          type: integer
        behavioralClusters:
          type: string
        referralRecommendations:
          type: string
        clinicalSummary:
          type: string

    # ─── Availability ────────────────────────────────────────────────────────
...
    TherapySessionNote:
      type: object
      required: [id, therapyPlanId, childId, sessionNumber, sessionDate, sessionType, createdAt, updatedAt]
      properties:
        id:
          type: integer
        therapyPlanId:
          type: integer
        childId:
          type: integer
        sessionNumber:
          type: integer
        sessionDate:
          type: string
        therapistName:
          type: ["string", "null"]
        sessionType:
          type: string
        durationMinutes:
          type: ["integer", "null"]
        goalsAddressed:
          type: ["string", "null"]
        activitiesPerformed:
          type: ["string", "null"]
        childResponse:
          type: ["string", "null"]
        progressObserved:
          type: ["string", "null"]
        challenges:
          type: ["string", "null"]
        parentFeedback:
          type: ["string", "null"]
        nextSessionPlan:
          type: ["string", "null"]
        overallRating:
          type: ["integer", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TherapySessionNoteInput:
      type: object
      required: [therapyPlanId, childId, sessionNumber, sessionDate]
      properties:
        therapyPlanId:
          type: integer
        childId:
          type: integer
        sessionNumber:
          type: integer
        sessionDate:
          type: string
        therapistName:
          type: string
        sessionType:
          type: string
        durationMinutes:
          type: integer
        goalsAddressed:
          type: string
        activitiesPerformed:
          type: string
        childResponse:
          type: string
        progressObserved:
          type: string
        challenges:
          type: string
        parentFeedback:
          type: string
        nextSessionPlan:
          type: string
        overallRating:
          type: integer
        createdBy:
          type: string
...