# lib/api-zod/src/generated/types/therapySessionNoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ACCENTECX AI CARE Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TherapySessionNoteInput {
  therapyPlanId: number;
  childId: number;
  sessionNumber: number;
  sessionDate: string;
  therapistName?: string;
  sessionType?: string;
  durationMinutes?: number;
  goalsAddressed?: string;
  activitiesPerformed?: string;
  childResponse?: string;
  progressObserved?: string;
  challenges?: string;
  parentFeedback?: string;
  nextSessionPlan?: string;
  overallRating?: number;
  createdBy?: string;
}